requirements.txt
pandas
yfinance
finnhub-python
python-dotenv
memory-profiler
jupyter
notebook
ipykernel
